InCoreIntegrals

Documentation for InCoreIntegrals.

InCoreIntegrals.InCoreInts

InCoreIntegrals.compute_energy

InCoreIntegrals.compute_energy

InCoreIntegrals.orbital_rotation

InCoreIntegrals.subset

InCoreIntegrals.subset

InCoreIntegrals.InCoreInts — Type

-h0::Real                # constant energy shift
-h1::Array{T,2}          # one electron integrals
-h2::Array{T,4}          # two electron integrals (chemist's notation)

Type to hold a second quantized Hamiltonian coefficients in memory

InCoreIntegrals.compute_energy — Method

compute_energy(h0, h1, h2, rdm1, rdm2)

Given an energy shift h0, 1e integrals h1, and 2e ints h2 along with a 1rdm and 2rdm on the same space, return the energy

InCoreIntegrals.compute_energy — Method

compute_energy(ints::InCoreInts, rdm1, rdm2)

Return energy defined by rdm1 and rdm2

InCoreIntegrals.orbital_rotation — Method

orbital_rotation(ints::InCoreInts, U)

Transform electronic integrals, by U i.e.,

\[h_{pq} = U_{rp}h_{rs}U_{sq}\]

\[(pq|rs) = (tu|vw)U_{tp}U_{uq}U_{vr}U_{ws}\]

InCoreIntegrals.subset — Method

subset(ints::InCoreInts, list; rmd1a, rdm1b)

Extract a subset of integrals acting on orbitals in list, returned as InCoreInts type and contract a 1rdm to give effectve 1 body interaction

Arguments

ints::InCoreInts: Integrals for full system

list: list of orbital indices in subset

rdm1a: 1RDM for embedding α density to make CASCI hamiltonian

rdm1b: 1RDM for embedding β density to make CASCI hamiltonian

InCoreIntegrals.subset — Method

subset(ints::InCoreInts, list)

Extract a subset of integrals acting on orbitals in list, returned as InCoreInts type. Because a constant is necessarily a global quantity, we return an InCoreInts object with h0 = 0.0.

Arguments

ints::InCoreInts: Integrals for full system

list: list of orbital indices in subset